refactor(runtime): Move `Bank` transaction methods to submodule
runtime/src/bank.rs is way too large.

Split transaction-related Bank methods out of runtime/src/bank.rs into
runtime/src/bank/transaction.rs. This covers transaction preparation,
locking, simulation, execution/commit, status lookup, verification, and
test/dev transaction helpers while preserving the existing Bank API.

Update bank tests to import transaction helper types directly.
